false is a valid datasource property
We ran into an issue where we noticed a datasource property not propagating to the hikari connections.  Specifically the property we needed to set was `useLegacyDatetimeCode`, and we needed to set it to `false`.

The (if value ...) was essentially throwing away our false setting instead of setting the property to false.  Our workaround was to do this in our hikari datasource spec:

{:use-legacy-datetime-code "false"}

Perhaps this is the recommended solution, and if so, we can edit the README to reflect that.  Currently the README shows properties being set with boolean values (not strings).